Ticket #231: Open stream buffer for XML parser (using fmemopen where available) align macro for CLANG compiler
git-svn-id: https://svn.code.sf.net/p/tcnopen/trdp/trunk@1831 3b5a3598-5f4e-4449-9e63-bd40438bfec0

+++ trdp/src/vos/common/vos_utils.c
@@ -363,46 +363,46 @@ static VOS_ERR_T vos_initRuntimeConsts (void)
 
     if (sAlignINT8 != ALIGNOF(INT8))
     {
-        vos_printLog(VOS_LOG_ERROR, "Unexpected alignment: %u != %u [ALIGNOF(INT8)]\n", sAlignINT8, ALIGNOF(INT8));
+        vos_printLog(VOS_LOG_ERROR, "Unexpected alignment: %u != %u [ALIGNOF(INT8)]\n", sAlignINT8, (unsigned int) ALIGNOF(INT8));
     }
     else if (sAlignINT16 != ALIGNOF(INT16))
     {
-        vos_printLog(VOS_LOG_ERROR, "Unexpected alignment: %u != %u [ALIGNOF(INT16)]\n", sAlignINT16, ALIGNOF(INT16));
+        vos_printLog(VOS_LOG_ERROR, "Unexpected alignment: %u != %u [ALIGNOF(INT16)]\n", sAlignINT16, (unsigned int) ALIGNOF(INT16));
     }
     else if (sAlignINT32 != ALIGNOF(INT32))
     {
-        vos_printLog(VOS_LOG_ERROR, "Unexpected alignment: %u != %u [ALIGNOF(INT32)]\n", sAlignINT32, ALIGNOF(INT32));
+        vos_printLog(VOS_LOG_ERROR, "Unexpected alignment: %u != %u [ALIGNOF(INT32)]\n", sAlignINT32, (unsigned int) ALIGNOF(INT32));
     }
     else if (sAlignREAL32 != ALIGNOF(REAL32))
     {
-        vos_printLog(VOS_LOG_ERROR, "Unexpected alignment: %u != %u [ALIGNOF(REAL32)]\n", sAlignREAL32, ALIGNOF(REAL32));
+        vos_printLog(VOS_LOG_ERROR, "Unexpected alignment: %u != %u [ALIGNOF(REAL32)]\n", sAlignREAL32, (unsigned int) ALIGNOF(REAL32));
     }
     else if (sAlignTIMEDATE48 != ALIGNOF(TIMEDATE48))
     {
         vos_printLog(VOS_LOG_ERROR, "Unexpected alignment: %u != %u [ALIGNOF(TIMEDATE48)]\n", sAlignTIMEDATE48,
-                     ALIGNOF(TIMEDATE48));
+                     (unsigned int) ALIGNOF(TIMEDATE48));
     }
     else if (sAlignINT64 != ALIGNOF(INT64))
     {
-        vos_printLog(VOS_LOG_ERROR, "Unexpected alignment: %u != %u [ALIGNOF(INT64)]\n", sAlignINT64, ALIGNOF(INT64));
+        vos_printLog(VOS_LOG_ERROR, "Unexpected alignment: %u != %u [ALIGNOF(INT64)]\n", sAlignINT64, (unsigned int) ALIGNOF(INT64));
     }
     else if (sAlignREAL64 != ALIGNOF(REAL64))
     {
-        vos_printLog(VOS_LOG_ERROR, "Unexpected alignment: %u != %u [ALIGNOF(REAL64)]\n", sAlignREAL64, ALIGNOF(REAL64));
+        vos_printLog(VOS_LOG_ERROR, "Unexpected alignment: %u != %u [ALIGNOF(REAL64)]\n", sAlignREAL64, (unsigned int) ALIGNOF(REAL64));
     }
     else if (sAlignTIMEDATE48Array1 != ALIGNOF(TIMEDATE48))
     {
         vos_printLog(VOS_LOG_ERROR,
                      "Unexpected alignment 1: %u != %u [ALIGNOF(TIMEDATE48)]\n",
                      sAlignTIMEDATE48Array1,
-                     ALIGNOF(TIMEDATE48));
+                     (unsigned int) ALIGNOF(TIMEDATE48));
     }
     else if (sAlignTIMEDATE48Array2 != ALIGNOF(TIMEDATE48))
     {
         vos_printLog(VOS_LOG_ERROR,
                      "Unexpected alignment 2: %u != %u [ALIGNOF(TIMEDATE48)]\n",
                      sAlignTIMEDATE48Array2,
-                     ALIGNOF(TIMEDATE48));
+                     (unsigned int) ALIGNOF(TIMEDATE48));
     }
     else
     {
